GCP/requirements.txt
torch==2.2.0
transformers==4.32.1
google-cloud-storage==1.16.1
sentencepiece==0.2.0